uploaded css, html, js files (Ari Li)
Changes made: (Also see comments in files)
js: added hideMsg, rewrote showMsg, add flags and timestamp to track the movement. Adopt a modal implementation and its functions. 
html: added class items for modal message box and initiation.
css: additions for modal message display

I noticed the alert box is not possible to be controlled and setTimeout for auto-close or fade-out. Hence, I adopted modal message box.
A thought of improvement is to have a sidebar/ panel to slide-show the message.