Exercise submission by bejumi
My goal was to meet the requirements of the exercise using existing,
freely available libraries. After examining the features and API for a
number of libraries, I settled on toastrjs as it was capable of meeting
all the requirements and uses jquery which the exercise already relied
on. After implementing the notification library, I went about solving a
pet peeve I have always had with notification systems. When a message is
dismissed (perhaps because I was in the middle of something or it popped
over something I was interacting with) there is often no way to review
that message later (I'm looking at you windows, android and ios). To fix
that, I decided to store a message history so the user could review any
missed messages later. I followed the same process in finding a suitable
library for handling storage. I started with basiljs but switched to
lockrjs for better handling of collections. See the README.md file for
more details.